dagster_card_processor/requirements.txt
dagster
dagster-webserver
dagster-dbt
google-generativeai
dbt-core
dbt-duckdb
python-dotenv
